refactor: normalize package names to match module buckets
Align Kotlin packages with the kotlin-conventions bucket structure so
the architecture layers map cleanly to ..common.. / ..modules..:
- com.tj.boilerplate.profile     → com.tj.boilerplate.common.profile
- com.tj.boilerplate.lock        → com.tj.boilerplate.modules.lock
- com.tj.boilerplate.pagination  → com.tj.boilerplate.modules.pagination

(storage.*, coreapi.*, modules.localcache already conformed.)

Lets the ArchUnit layeredArchitecture define Modules/Common with single
..modules.. / ..common.. identifiers instead of enumerating each module
package — new common/modules modules are covered automatically.

@SpringBootApplication classes scan from the root package, so deeper
packages stay in scan scope; no string-literal package references to
update. Verified: ./gradlew clean build → SUCCESSFUL.
