+++ common/profile/src/main/kotlin/com/tj/boilerplate/common/profile/DevProfile.kt
@@ -1,7 +1,7 @@
-package com.tj.boilerplate.profile
+package com.tj.boilerplate.common.profile
 
 import org.springframework.context.annotation.Profile
 
-@Profile("live")
+@Profile("dev")
 @Target(AnnotationTarget.CLASS)
-annotation class LiveProfile
+annotation class DevProfile
