+++ common/profile/src/main/kotlin/com/tj/boilerplate/common/profile/LocalProfile.kt
@@ -1,4 +1,4 @@
-package com.tj.boilerplate.profile
+package com.tj.boilerplate.common.profile
 
 import org.springframework.context.annotation.Profile
