+++ common/profile/src/test/kotlin/com/tj/boilerplate/common/profile/ProfileCheckerTest.kt
@@ -1,4 +1,4 @@
-package com.tj.boilerplate.profile
+package com.tj.boilerplate.common.profile
 
 import io.mockk.every
 import io.mockk.mockk
