+++ modules/lock/src/main/kotlin/com/tj/boilerplate/modules/lock/EmbeddedRedisLockConfiguration.kt
@@ -1,6 +1,6 @@
-package com.tj.boilerplate.lock
+package com.tj.boilerplate.modules.lock
 
-import com.tj.boilerplate.profile.LocalProfile
+import com.tj.boilerplate.common.profile.LocalProfile
 import jakarta.annotation.PostConstruct
 import jakarta.annotation.PreDestroy
 import org.redisson.Redisson
