+++ modules/lock/src/main/kotlin/com/tj/boilerplate/modules/lock/RedisLockProperties.kt
@@ -1,7 +1,7 @@
-package com.tj.boilerplate.lock
+package com.tj.boilerplate.modules.lock
 
-import com.tj.boilerplate.profile.DevProfile
-import com.tj.boilerplate.profile.LiveProfile
+import com.tj.boilerplate.common.profile.DevProfile
+import com.tj.boilerplate.common.profile.LiveProfile
 import org.springframework.boot.context.properties.ConfigurationProperties
 
 @DevProfile
