+++ modules/pagination/src/main/kotlin/com/tj/boilerplate/modules/pagination/CursorPaginationParam.kt
@@ -1,4 +1,4 @@
-package com.tj.boilerplate.pagination
+package com.tj.boilerplate.modules.pagination
 
 data class CursorPaginationParam(
     val size: Int,
