+++ storage/redis/src/main/kotlin/com/tj/boilerplate/storage/redis/configuration/RedisStorageConfiguration.kt
@@ -1,7 +1,7 @@
 package com.tj.boilerplate.storage.redis.configuration
 
-import com.tj.boilerplate.profile.DevProfile
-import com.tj.boilerplate.profile.LiveProfile
+import com.tj.boilerplate.common.profile.DevProfile
+import com.tj.boilerplate.common.profile.LiveProfile
 import org.springframework.boot.autoconfigure.data.redis.RedisProperties
 import org.springframework.context.annotation.Bean
 import org.springframework.context.annotation.Configuration
